server: Fix panic with too many users
With too many users, the query had a giant filter `IN (u1, u2, u3,
...)`. This would cause a panic below if the number of users was too
high. In PostgreSQL, we can pass the users as an array instead, but that
doesn't work with SQLite. Instead, we repeat the filter from the
previous query to get the same users/groups, as a subquery.